requirements.txt
requests>=2.31.0
pypdf>=4.2.0
python-dotenv>=1.0.1
PyJWT>=2.8.0
cryptography>=42.0.0
matplotlib>=3.8.0
